pyproject.toml
[project]
name = "mcp-todoaist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "caldav>=2.0.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.1",
    "vobject>=0.9.9",
]
